fix: Errors in AWS Tests (#3523)
* fix: error around retentionperiod

* chore: debug what is happening on the server

* fix: trace not visible

* fix: skip retention period

* fix: date comparison issues; mixed pipeline isssues

* fix: issues with enumerated type for FIFO queues

* fix: splitting up queue and topic for cleaner code now we support point-to-point and pub-sub

* fix: rename Subscription identifer to subscriptionName for clarity

* fix: fixing Aws tests

* fix: make sqs vs sns more explicit

* fix: separate sqs and sns subscriptions/publications

* fix: time formats are not correct following change to TimeSpan.

* fix: ensure sensible defaults when we create an SqsPublication

* fix: we need to supply a channel name on an SqsProducer if we intend to search by channel name

* fix: ensure we don't keep redeterming the Url

* fix: common attributes method must check if queue is DLQ or normal

* fix: consistency across pump type, async, and subscription property changes

* fix: issues with dictionary keys not matching

* chore: flow proactor edits to reactor

* fix: disparities in create channel methods leading to sync tests failing

* fix: supply topicattributes to subscription for Fifo

* fix: Fifo PubSub tests fixed with shared topic attributes

* fix: add the channelname to the SqsSubscription.cs

* fix: update publish to use channel name

* fix: use channelname in the sqs publication

* fix: sqs required the channel name

* fix: ensure sqs fifo tests correct

* fix: update Fifo SQS tests to correct parameters

* chore: clean up names, remove some fragile markers and see if they now work on CI

* fix: assume should still run validation

* fix:re-enable DLQ tests

* fix: async test requires async message mapper

* fix: re-drive policy should take channel name

* fix: avoid having both queueurl and channelname, just make the channelname stand for both, and use the findby to understand it.

* fix: correct configuration within tests

* chore: move scheduler tests to own repo

* chore: allow ci of awsscheduler tests

* update adrs for scheduling and cloudevents

* chore merge Use source generated logging (#3579)

* feature: Use source generated logging

* Remove newlines

* Revert AWSMessagingGateway

* Revert AWSMessagingGateway

* Revert MQTTMessageProducer

* Revert MQTTMessageProducer

* Fix and add adrs

* chore: merge mqtt changes

* [Fix] Scheduler Tests were breaking (#3582)

* update packages

* fix: we need to stop outbox entries dissappearing during long-running scheduler tasks

* feature: refactoring of MQTT Support to use 4.3 MQTTnet (#3578)

* Refactoring of MQTT Support to use 4.3 MQTTnet

* Removed MQTTnet v5 support for UserProperties.  Removed Comments.

* Fix logging of Exception when Test context is not available

* Increase wait to 1 sec before testing for number of messages

* SpinUntil instead of trying to time when requests will be processed

* Update target platforms for test helpers to net8.0;net9.0

* Adjust timeout to see if all messages are processed.

* Reworking logging to reduce logging for routine message delivery

* Fixed static Log class merge errors

* Fix up passing of exception and port to new Log API.

---------


* chore: merge issues

* fix: remove version numbers from dependencies for centralized package management

* fix: typo in path to scheduler tests.

---------

+++ tests/Paramore.Brighter.AWS.Tests/MessagingGateway/Sns/Fifo/Reactor/When_queues_missing_verify_throws.cs
@@ -9,26 +9,28 @@ using Xunit;
 namespace Paramore.Brighter.AWS.Tests.MessagingGateway.Sns.Fifo.Reactor;
 
 [Trait("Category", "AWS")]
-public class AWSValidateQueuesTests : IDisposable, IAsyncDisposable
+public class AwsValidateQueuesTests : IDisposable, IAsyncDisposable
 {
     private readonly AWSMessagingGatewayConnection _awsConnection;
     private readonly SqsSubscription<MyCommand> _subscription;
-    private ChannelFactory _channelFactory;
+    private ChannelFactory? _channelFactory;
 
-    public AWSValidateQueuesTests()
+    public AwsValidateQueuesTests()
     {
         var channelName = $"Producer-Send-Tests-{Guid.NewGuid().ToString()}".Truncate(45);
         string topicName = $"Producer-Send-Tests-{Guid.NewGuid().ToString()}".Truncate(45);
         var routingKey = new RoutingKey(topicName);
+        var topicAttributes = new SnsAttributes { Type = SqsType.Fifo };
 
         _subscription = new SqsSubscription<MyCommand>(
-            name: new SubscriptionName(channelName),
+            subscriptionName: new SubscriptionName(channelName),
             channelName: new ChannelName(channelName),
+            channelType: ChannelType.PubSub,
             routingKey: routingKey,
-            messagePumpType: MessagePumpType.Reactor,
-            makeChannels: OnMissingChannel.Validate,
-            sqsType: SnsSqsType.Fifo
-        );
+            queueAttributes: new SqsAttributes(type: SqsType.Fifo), 
+            topicAttributes: topicAttributes,
+            messagePumpType: MessagePumpType.Reactor, 
+            makeChannels: OnMissingChannel.Validate);
 
         _awsConnection = GatewayFactory.CreateFactory();
 
@@ -36,7 +38,7 @@ public class AWSValidateQueuesTests : IDisposable, IAsyncDisposable
         var producer = new SnsMessageProducer(_awsConnection,
             new SnsPublication
             {
-                MakeChannels = OnMissingChannel.Create, SnsAttributes = new SnsAttributes { Type = SnsSqsType.Fifo }
+                MakeChannels = OnMissingChannel.Create, TopicAttributes = topicAttributes
             });
         producer.ConfirmTopicExistsAsync(topicName).Wait();
     }
@@ -52,7 +54,8 @@ public class AWSValidateQueuesTests : IDisposable, IAsyncDisposable
 
     public void Dispose()
     {
-        _channelFactory.DeleteTopicAsync().Wait();
+        if (_channelFactory != null)
+            _channelFactory.DeleteTopicAsync().Wait();
     }
 
     public async ValueTask DisposeAsync()
