+++ tests/Paramore.Brighter.AWS.Tests/MessagingGateway/Sns/Standard/Reactor/When_queues_missing_verify_throws.cs
@@ -9,21 +9,22 @@ using Xunit;
 namespace Paramore.Brighter.AWS.Tests.MessagingGateway.Sns.Standard.Reactor;
 
 [Trait("Category", "AWS")] 
-public class AWSValidateQueuesTests  : IDisposable, IAsyncDisposable
+public class AwsValidateQueuesTests  : IDisposable, IAsyncDisposable
 {
     private readonly AWSMessagingGatewayConnection _awsConnection;
     private readonly SqsSubscription<MyCommand> _subscription;
-    private ChannelFactory _channelFactory;
+    private ChannelFactory? _channelFactory;
 
-    public AWSValidateQueuesTests()
+    public AwsValidateQueuesTests()
     {
         var channelName = $"Producer-Send-Tests-{Guid.NewGuid().ToString()}".Truncate(45);
         string topicName = $"Producer-Send-Tests-{Guid.NewGuid().ToString()}".Truncate(45);
         var routingKey = new RoutingKey(topicName);
             
         _subscription = new SqsSubscription<MyCommand>(
-            name: new SubscriptionName(channelName),
+            subscriptionName: new SubscriptionName(channelName),
             channelName: new ChannelName(channelName),
+            channelType: ChannelType.PubSub,
             routingKey: routingKey,
             messagePumpType: MessagePumpType.Reactor,
             makeChannels: OnMissingChannel.Validate
@@ -52,11 +53,13 @@ public class AWSValidateQueuesTests  : IDisposable, IAsyncDisposable
  
     public void Dispose()
     {
-        _channelFactory.DeleteTopicAsync().Wait(); 
+        if (_channelFactory != null)
+            _channelFactory.DeleteTopicAsync().Wait(); 
     }
         
     public async ValueTask DisposeAsync()
     {
-        await _channelFactory.DeleteTopicAsync(); 
+        if (_channelFactory != null)
+            await _channelFactory.DeleteTopicAsync(); 
     }
 }
